requirements.txt
asgiref==3.11.1
Django==4.2.28
et_xmlfile==2.0.0
openpyxl==3.1.5
pillow==10.4.0
sqlparse==0.5.5
